refactor: dictionaryplants hook으로 분리
